requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.2
aiosignal==1.3.2
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
attrs==23.2.0
Automat==22.10.0
bcc==0.29.1
bcrypt==3.2.2
beautifulsoup4==4.13.4
blinker==1.7.0
Brotli==1.1.0
cachetools==5.5.2
certifi==2023.11.17
cffi==1.17.1
chardet==5.2.0
click==8.1.6
colorama==0.4.6
command-not-found==0.3
configobj==5.0.8
constantly==23.10.4
cryptography==41.0.7
cssselect2==0.8.0
dbus-python==1.3.2
distro==1.9.0
distro-info==1.7+build1
docker==5.0.3
docker-compose==1.29.2
dockerpty==0.4.1
docopt==0.6.2
fastapi==0.115.12
fonttools==4.58.0
frozenlist==1.6.0
gitdb==4.0.12
GitPython==3.1.44
h11==0.16.0
httpcore==1.0.9
httplib2==0.20.4
httpx==0.28.1
hyperlink==21.0.0
idna==3.6
incremental==22.10.0
Jinja2==3.1.6
jiter==0.10.0
jsonpointer==2.0
jsonschema==4.10.3
launchpadlib==1.11.0
lazr.restfulclient==0.14.6
lazr.uri==1.0.6
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
multidict==6.4.4
narwhals==1.41.0
netaddr==0.8.0
netifaces==0.11.0
numpy==2.2.6
oauthlib==3.2.2
openai==1.82.0
packaging==24.0
pandas==2.2.3
pexpect==4.9.0
pillow==10.4.0
propcache==0.3.1
protobuf==4.25.8
ptyprocess==0.7.0
pyarrow==20.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.22
pydantic==2.11.5
pydantic_core==2.33.2
pydeck==0.9.1
pydyf==0.11.0
Pygments==2.17.2
PyGObject==3.48.2
PyHamcrest==2.1.0
PyJWT==2.7.0
pyOpenSSL==23.2.0
pyparsing==3.1.1
pyphen==0.17.2
pyrsistent==0.20.0
pyserial==3.5
python-apt==2.7.7+ubuntu3
python-dateutil==2.9.0.post0
python-debian==0.1.49+ubuntu2
python-dotenv==1.0.1
python-magic==0.4.27
python-multipart==0.0.6
pytz==2025.2
PyYAML==6.0.1
requests==2.31.0
rich==13.7.1
service-identity==24.1.0
setuptools==68.1.2
six==1.16.0
smmap==5.0.2
sniffio==1.3.1
sos==4.5.6
soupsieve==2.7
ssh-import-id==5.11
starlette==0.46.2
streamlit==1.45.1
systemd-python==235
tenacity==8.5.0
texttable==1.6.7
tinycss2==1.4.0
tinyhtml5==2.0.0
toml==0.10.2
tornado==6.5.1
tqdm==4.67.1
Twisted==24.3.0
typing-inspection==0.4.1
typing_extensions==4.13.2
tzdata==2025.2
tzlocal==5.3.1
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufw==0.36.2
urllib3==2.0.7
uvicorn==0.24.0
validators==0.35.0
wadllib==1.3.6
watchdog==6.0.0
weasyprint==65.1
webencodings==0.5.1
websocket-client==1.7.0
wheel==0.42.0
xkit==0.0.0
yarl==1.20.0
zipp==3.22.0
zope.interface==6.1
zopfli==0.2.3.post1